+++ src/components/POILayers/index.js
@@ -3,3 +3,4 @@ export { default as MallsLayer } from './MallsLayer';
 export { default as FitnessLayer } from './FitnessLayer';
 export { default as SupermarketsLayer } from './SupermarketsLayer';
 export { default as KyivstarLayer } from './KyivstarLayer';
+export { default as ApolloClubsLayer } from './ApolloClubsLayer';